Move set image animated method to helper extension file
