+++ HotelSearch/HotelSearchiOS/View/HotelCell.swift
@@ -60,17 +60,3 @@ private extension HotelCell {
     }
     
 }
-
-extension UIImageView {
-    func setImageAnimated(_ newImage: UIImage?) {
-        image = newImage
-        
-        guard newImage != nil else { return }
-        
-        alpha = 0
-        UIView.animate(withDuration: 0.25) {
-            self.alpha = 1
-        }
-    }
-}
-
